Inclusão de mais uma exceção e correções gerais

+++ MVC/controle/controlador_cliente.py
@@ -4,6 +4,7 @@ from MVC.entidade.cliente import Cliente
 from MVC.exceptions.botaoErradoException import BotaoErradoException
 from MVC.exceptions.jaExistenteException import JaExistenteException
 from MVC.exceptions.listaVaziaException import ListaVaziaException
+from MVC.exceptions.nadaSelecionadoException import NadaSelecionadoException
 from MVC.persistencia.cliente_dao import ClienteDAO
 
 
@@ -45,31 +46,36 @@ class ControladorCliente:
 
     def incluir_cliente(self, values=None):
         while True:
-            botao, dados_cliente = self.__tela_dados_cliente.open(
-                dados_cliente={"codigo": "", "nome": "", "pais_origem": "",
-                               "local_sede": ""})
+            botao, dados_cliente = self.__tela_dados_cliente.open(dados_cliente={"codigo": "", "nome": "",
+                                                                                 "pais_origem": "", "local_sede": ""})
 
             self.__tela_dados_cliente.close()
 
             try:
-                int(dados_cliente["codigo"])
+                dados_cliente['codigo'] = int(dados_cliente["codigo"])
                 if dados_cliente == {"codigo": "", "nome": "", "pais_origem": "", "local_sede": ""} or \
                         (dados_cliente["nome"]).isdigit() == True or (dados_cliente["pais_origem"]).isdigit() == True \
                         or (dados_cliente["local_sede"]).isdigit() == True:
                     raise ValueError
+
                 for cliente in self.__cliente_dao.get_all():
-                    if cliente.nome == dados_cliente['nome']:
+                    if cliente.nome == dados_cliente['nome'] or cliente.codigo == dados_cliente['codigo']:
                         raise JaExistenteException
+
                 cliente = Cliente(dados_cliente["nome"], dados_cliente["pais_origem"], dados_cliente["local_sede"],
-                                  int(dados_cliente["codigo"]))
+                                  dados_cliente["codigo"])
                 self.__cliente_dao.persist(cliente)
                 self.__tela_cliente_gui.close()
                 break
+
             except ValueError:
                 self.__tela_cliente_gui.show_message('Atenção', 'Código inválido, tente novamente!')
+                self.__tela_cliente_gui.close()
                 continue
             except JaExistenteException:
                 self.__tela_cliente_gui.show_message('Atenção', 'Cliente já existente, tente novamente!')
+                self.__tela_cliente_gui.close()
+                continue
 
     def monta_dict_clientes(self):
         clientes = []
@@ -83,6 +89,8 @@ class ControladorCliente:
             if checagem_cliente == 0:
                 self.__tela_cliente_gui.close()
                 raise ListaVaziaException
+            if dados_cliente['lb_itens'] == []:
+                raise NadaSelecionadoException
             else:
                 nome_cliente_alterado = dados_cliente['lb_itens'][0]
 
@@ -91,6 +99,8 @@ class ControladorCliente:
                         dados_cliente = {"codigo": cliente.codigo, "nome": cliente.nome,
                                          "pais_origem": cliente.pais_origem,
                                          "local_sede": cliente.local_sede}
+
+                codigo_cliente_alterado = dados_cliente['codigo']
 
                 while True:
                     try:
@@ -102,25 +112,35 @@ class ControladorCliente:
                                 or (dados_cliente["local_sede"]).isdigit() == True:
                             raise ValueError
 
-                        lista_clientes = self.__cliente_dao.get_all()
+                        for cliente in self.__cliente_dao.get_all():
+                            if cliente.nome == values['nome'] or cliente.codigo == values['codigo']:
+                                raise JaExistenteException
 
-                        for cliente in lista_clientes:
+                        for cliente in self.__cliente_dao.get_all():
                             if cliente.nome == nome_cliente_alterado:
-                                cliente.nome = values["nome"]
-                                cliente.pais_origem = values["pais_origem"]
-                                cliente.local_sede = values["local_sede"]
-                                cliente.codigo = values["codigo"]
-                                print(values['codigo'])
-                                print('Nome: ', cliente.nome,
-                                      '\nPais_origem: ', cliente.pais_origem,
-                                      '\nLocal_sede: ', cliente.local_sede,
-                                      '\nCódigo: ', cliente.codigo
-                                      )  # CÓDIGO E CHAVE NÃO ESTÃO MUDANDO
+                                cliente_desejado = cliente
 
-                                self.__cliente_dao.persist(cliente)  # IMPORTANTE
+                        if values['codigo'] != codigo_cliente_alterado:
+                            self.__cliente_dao.remove(cliente_desejado)
+                            cliente = Cliente(values["nome"], values["pais_origem"], values["local_sede"],
+                                              values["codigo"])
+                            self.__cliente_dao.persist(cliente)
+                            self.__tela_cliente_gui.close()
+
+                        elif values['codigo'] == codigo_cliente_alterado:
+                            cliente_desejado.nome = values["nome"]
+                            cliente_desejado.pais_origem = values["pais_origem"]
+                            cliente_desejado.local_sede = values["local_sede"]
+
+                        # se o código for diferente do anterior, exclui da cache e adiciona o novo
+
                         break
+
                     except ValueError:
                         self.__tela_cliente_gui.show_message('Atenção', 'Código inválido, tente novamente!')
+                        continue
+                    except JaExistenteException:
+                        self.__tela_cliente_gui.show_message('Atenção', 'Cliente já existente, tente novamente!')
                         continue
 
                 self.__tela_dados_cliente.close()
@@ -128,6 +148,10 @@ class ControladorCliente:
 
         except ListaVaziaException:
             self.__tela_cliente_gui.show_message("Atenção!", "Ainda não há clientes cadastrados.")
+            self.__tela_cliente_gui.close()
+        except NadaSelecionadoException:
+            self.__tela_cliente_gui.show_message("Ops!", "Nenhum cliente foi selecionado.")
+            self.__tela_cliente_gui.close()
 
     def checar_lista_clientes(self):
         clientes = []
@@ -137,34 +161,58 @@ class ControladorCliente:
             return 0
 
     def excluir_cliente(self, dados_cliente):
-        if self.__cliente_dao.get_all() == []:
+        try:
+            checagem_cliente = self.checar_lista_clientes()
+            if checagem_cliente == 0:
+                self.__tela_cliente_gui.close()
+                raise ListaVaziaException
+            if dados_cliente['lb_itens'] == []:
+                raise NadaSelecionadoException
+            else:
+                nome_cliente_excluido = dados_cliente['lb_itens'][0]
+
+                lista_clientes = self.__cliente_dao.get_all()
+                cliente_excluido = ''
+                for cliente in lista_clientes:
+                    if cliente.nome == nome_cliente_excluido:
+                        cliente_excluido = cliente
+                self.__cliente_dao.remove(cliente_excluido)
+                del cliente_excluido
+                self.__tela_cliente_gui.close()
+        except ListaVaziaException:
             self.__tela_cliente_gui.show_message("Atenção!", "Ainda não há clientes cadastrados.")
-
-        nome_cliente_excluido = dados_cliente['lb_itens'][0]
-
-        lista_clientes = self.__cliente_dao.get_all()
-        cliente_excluido = ''
-        for cliente in lista_clientes:
-            if cliente.nome == nome_cliente_excluido:
-                cliente_excluido = cliente
-        self.__cliente_dao.remove(cliente_excluido)
-        del cliente_excluido
-        self.__tela_cliente_gui.close()
+            self.__tela_cliente_gui.close()
+        except NadaSelecionadoException:
+            self.__tela_cliente_gui.show_message("Ops!", "Nenhum cliente foi selecionado.")
+            self.__tela_cliente_gui.close()
 
     def mostrar_detalhes(self, values):
-        nome_desejado = values['lb_itens'][0]
-        for cliente in self.__cliente_dao.get_all():
-            if cliente.nome == nome_desejado:
-                pais_origem = cliente.pais_origem
-                local_sede = cliente.local_sede
-                codigo = cliente.codigo
+        try:
+            checagem_cliente = self.checar_lista_clientes()
+            if checagem_cliente == 0:
+                self.__tela_cliente_gui.close()
+                raise ListaVaziaException
+            if values['lb_itens'] == []:
+                raise NadaSelecionadoException
+            else:
+                nome_desejado = values['lb_itens'][0]
+                for cliente in self.__cliente_dao.get_all():
+                    if cliente.nome == nome_desejado:
+                        pais_origem = cliente.pais_origem
+                        local_sede = cliente.local_sede
+                        codigo = cliente.codigo
 
-        self.__tela_cliente_gui.close()
-        self.__tela_cliente_gui.show_message("Detalhes do Cliente:",
-                                             f'Nome: {nome_desejado}, País de origem: {pais_origem},  Local da sede: {local_sede}, Código: {codigo}.')
+                self.__tela_cliente_gui.close()
+                self.__tela_cliente_gui.show_message("Detalhes do Cliente:",
+                                                     f'Nome: {nome_desejado}, País de origem: {pais_origem}, Local da sede: {local_sede}, Código: {codigo}.')
+        except ListaVaziaException:
+            self.__tela_cliente_gui.show_message("Atenção!", "Ainda não há clientes cadastrados.")
+            self.__tela_cliente_gui.close()
+        except NadaSelecionadoException:
+            self.__tela_cliente_gui.show_message("Ops!", "Nenhum cliente foi selecionado.")
+            self.__tela_cliente_gui.close()
 
     def retornar(self, values):
-        self.__tela_cliente_gui.close()
         self.__controlador_sistema.abre_tela()
 
     def abre_tela(self):
@@ -176,7 +224,22 @@ class ControladorCliente:
         continua = True
         while continua:
             dict_clientes = self.monta_dict_clientes()
-            button, values = self.__tela_cliente_gui.open(dict_clientes)
+
+            while True:
+                button, values = self.__tela_cliente_gui.open(dict_clientes)
+
+                try:
+                    if button == 'Incluir em Missão' or button == 'Incluir mais Clientes':
+                        raise BotaoErradoException
+                    break
+                except BotaoErradoException:
+                    self.__tela_cliente_gui.show_message('Ops!',
+                                                         'Você só deve clicar nesse botão quando estiver cadastrando '
+                                                         'uma missão!')
+                    self.__tela_cliente_gui.close()
+                    continue
+
+            self.__tela_cliente_gui.close()
             lista_opcoes[button](values)
 
             # lista_opcoes[self.__tela_cliente_gui.open(self.monta_dict_clientes())]()
